Deleted old auton enums
