+++ common/src/main/java/us/ilite/common/types/auton/EStartingPosition.java
@@ -1,1 +1,14 @@
 package us.ilite.common.types.auton;
+
+public enum EStartingPosition {
+    MIDDLE_LEFT,
+	MIDDLE_RIGHT,
+    BOTTOM_LEFT,
+    BOTTOM_RIGHT,
+    UNKNOWN;
+  public static EStartingPosition intToEnum(int num) {
+    if(num == -1)return UNKNOWN;
+    if(num >= values().length || num < 0) return UNKNOWN;
+    return values()[num];
+  }
+}
